Remove oldparsers (#161)
